rake added and edits made

+++ app/controllers/items_controller.rb
@@ -22,17 +22,19 @@ class ItemsController < ApplicationController
   end
 
   def destroy
-    @user = current_user
-    @item = @user.items.find(params[:id])
-    if @item.destroy
-      flash[:notice] = "Item marked complete!"
-    else
-      flash[:alert] = "Item couldn't be marked complete. Please try again."
+    @user = User.find(params[:user_id])
+    if @user == current_user
+      @item = Item.find(params[:id])
+      if @item.destroy
+        flash[:notice] = "Item marked complete!"
+      else
+        flash[:alert] = "Item couldn't be marked complete. Please try again."
+      end
     end
 
     respond_to do |format|
       format.html
-      format.js
+      format.js {}
     end
   end
